added signed apk with background playback working
I was (understandably) not allowed to post this in the Google Play Store due to the fact that this version uses GET_SERVICE in the app manifest to check if the BitChute is running in the foreground or background.  You must enable this for the app to work in background.  You can also download and compile to get a better idea of how I did it.

Manually download and install this APK if you're looking to have background playback in the app.